apps/api/pyproject.toml
[project]
name = "foodie-api"
version = "0.1.0"
description = "Foodie API"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "psycopg[binary]>=3.1",
  "sqlalchemy[asyncio]>=2.0",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "argon2-cffi>=23.1",
  "pyjwt>=2.8",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "httpx>=0.27",
  "aiosqlite>=0.20",
]

[tool.uvicorn]
factory = false
